Move consumer to models
