+++ source/components/GlobalNavigation/components/Notifications/ListContainer.js
@@ -1,7 +1,7 @@
 import React, { useContext } from 'react';
 import PropTypes from 'prop-types';
 
-import NotificationsContext from '../../utils/NotificationContext';
+import NotificationsContext from '../../models/NotificationContext';
 
 import MarkAllButton from './MarkAllButton';
 import List from './List';
