+++ source/components/GlobalNavigation/components/User/MobileLoggedIn.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import Button from '../../../Button';
 import Avatar from '../../../Avatar';
 
-import NotificationsContext from '../../utils/NotificationContext';
+import NotificationsContext from '../../models/NotificationContext';
 
 import UserModal from './UserModal';
